==> 2.Build Ratings Matrix.ipynb <==
# Building a Ratings Matrix

The purpose of this workbook is to build a user-item ratings matrix which will be used for a collaborative based recommender system.

The final output will be a ratings matrix which will be saved as a .pkl (pickle) file to be used in future workbooks.

***

import numpy as np
import pandas as pd
import pickle

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

business = pd.read_pickle('./Dataset/business.pkl')
review = pd.read_pickle('./Dataset/review_new.pkl')

business.shape
business.head(3)

review.shape
review.head(3)

Make sure there are no records of missing star ratings in the comment form.

len(review) - review['stars'].isna().value_counts()

Give each user a unique 'user_id' value. And save it as a new file.

user_table = review['user_id'].unique()
user_table = pd.DataFrame(user_table)
user_table = user_table.reset_index().rename(columns = {'index':'user_num',0:'user_id'})
user_table

pickle.dump(user_table, file=open('./Dataset/unique_user.pkl', 'wb+'))

user_table.shape
user_table.head()

business_table = review['business_id'].unique()
business_table = pd.DataFrame(business_table)
business_table = business_table.reset_index().rename(columns = {'index':'business_num',0:'business_id'})
business_table

pickle.dump(business_table, file=open('./Dataset/unique_business.pkl', 'wb+'))

business_table.shape
business_table.head()

num_user = len(user_table)
num_business = len(business_table)

print('unique users:', num_user, '\nunique businesses:', num_business)

%%time
ratings_array = np.full((num_user, num_business), np.nan)
ratings_array

Sort data frame by most recent to least recent date (in order to assist with dropping duplicates)

review_rate = review.loc[:,['user_id','business_id','stars','date']].copy()
review_rate.sort_values(by='date', ascending = False, inplace = True)

review_rate.shape
review_rate.head(3)

Only keep the most recent review.

review_rate.drop_duplicates(['user_id','business_id'], keep='first', inplace = True)
review_rate.shape
review_rate.head(3)

# Lookup the unique integer based index values for each unique user_id and business_id
review_rate = pd.merge(review_rate, user_table, how = 'left')
review_rate = pd.merge(review_rate, business_table, how = 'left')
review_rate.head(3)

Confirm that there are no missing values

review_rate['business_num'].isna().value_counts()
review_rate['user_num'].isna().value_counts()

# Transfer the star rating from the 'review_rate' df into the 'ratings_array' df
for row in review_rate.itertuples():
    user_index = row[5]
    business_index = row[6]
    rating_value = row[3]
    ratings_array[user_index, business_index] = rating_value

# Transform the 'ratings_array' into a data-frame
ratings_matrix = pd.DataFrame(ratings_array)
ratings_matrix

pickle.dump(ratings_matrix, file=open('./Dataset/ratings_matrix_new.pkl', 'wb+'))
