==> 3.Model Evaluation.ipynb <==
# Model Evaluation 
This book is used to build and hybrid user-based and item-based CF and assign weights to both by calculating the Root Mean Square Error (RMSE). In this way, we can make our prediction more precisely.

import numpy as np
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.model_selection import train_test_split

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

ratings_matrix = pd.read_pickle('./Dataset/ratings_matrix_new.pkl')
review = pd.read_pickle('./Dataset/review_new.pkl')
user_index = pd.read_pickle('./Dataset/unique_user.pkl')
business_index = pd.read_pickle('./Dataset/unique_business.pkl')

business_index
user_index

ratings_matrix
review

ratings_matrix.shape

# Item_based CF

def find_business_similarity(businessA, businessB, ratings_matrix):
    
    users_rated_businessA = ~ratings_matrix.loc[:, businessA].isna()
    users_rated_businessB = ~ratings_matrix.loc[:, businessB].isna()

    users_rated_AB = users_rated_businessA & users_rated_businessB

    ratings_of_businessA = ratings_matrix.loc[users_rated_AB, businessA].values.reshape(1, -1)
    ratings_of_businessB = ratings_matrix.loc[users_rated_AB, businessB].values.reshape(1, -1)
    
    similarity = cosine_similarity(ratings_of_businessA, ratings_of_businessB)[0][0]

    return similarity

def item_item_rating_prediction(target_user, target_business, ratings_matrix):

    similarities_to_target_business = []
    ratings_given_by_target_user = []
    
    list_of_businesses_rated_by_target_user = list(ratings_matrix.loc[:, ~ratings_matrix.iloc[target_user, :].isna()].columns)
    
    for other_business in list_of_businesses_rated_by_target_user:
        try:
            similarity = find_business_similarity(target_business, other_business, ratings_matrix)
            similarities_to_target_business.append(similarity)
            ratings_given_by_target_user.append(ratings_matrix.loc[target_user, other_business])
        except:
            pass
        
    return np.dot(ratings_given_by_target_user, similarities_to_target_business)/np.sum(similarities_to_target_business)

def find_user_similarity(userA, userB, ratings_matrix):
    
    businesses_rated_by_userA = ~ratings_matrix.loc[userA, :].isna()
    businesses_rated_by_userB = ~ratings_matrix.loc[userB, :].isna()
    
    businesses_rated_by_AB = businesses_rated_by_userA & businesses_rated_by_userB
    
    ratings_of_userA = ratings_matrix.loc[userA, businesses_rated_by_AB].values.reshape(1, -1)
    ratings_of_userB = ratings_matrix.loc[userB, businesses_rated_by_AB].values.reshape(1, -1)
    
    similarity = cosine_similarity(ratings_of_userA, ratings_of_userB)[0][0]
    
    return similarity

# User_based CF

def user_item_rating_prediction(target_user, target_business, ratings_matrix):
   
    similarities_to_target_user = []
    ratings_given_to_target_business = []
    
    list_of_users_rating_target_business = list(ratings_matrix[~ratings_matrix.iloc[:, target_business].isna()].index)
    
    for other_user in list_of_users_rating_target_business:
        try:
            similarity = find_user_similarity(target_user, other_user, ratings_matrix)
            similarities_to_target_user.append(similarity)
            ratings_given_to_target_business.append(ratings_matrix.loc[other_user, target_business])
        except:
            pass
    return np.dot(ratings_given_to_target_business, similarities_to_target_user)/np.sum(similarities_to_target_user)

# Split as train/test group

train_df, test_df = train_test_split(ratings_matrix, test_size = 0.00020)
train_df.shape
test_df.shape

test_users = pd.Series(list(test_df.index), name = 'user_num')
test_users

review_edit = review.copy()
review_edit = pd.merge(review_edit,user_index,how = 'left')
review_edit = pd.merge(review_edit,business_index,how = 'left')
review_edit.head(1)

review_edit = review_edit.sort_values('date', ascending = False)

review_edit = review_edit.drop_duplicates(subset = ['user_num','business_num'],keep='first')
review_edit.shape

review_edit = review_edit[['user_num', 'business_num', 'stars']]
review_edit.tail(1)

review_edit = pd.merge(review_edit, test_users, on = 'user_num', how = 'inner')
review_edit

review_edit.insert(3, 'check', 0)

review_edit

for record in review_edit.index:
    review_edit.loc[record, ['check']] = ratings_matrix.loc[review_edit.loc[record, ['user_num'][0]], review_edit.loc[record, ['business_num'][0]]]

(review_edit['stars'] - review_edit['check']).value_counts()

review_edit = review_edit.drop(columns = {'check'})

review_edit.insert(3, 'item-item', np.nan)
review_edit.insert(4, 'user-item', np.nan)
review_edit

%%time

for record in review_edit.index:
    user_num = review_edit.loc[record, ['user_num'][0]]
    business_num = review_edit.loc[record, ['business_num'][0]]
    
    review_edit.loc[record, ['user-item']] = round(user_item_rating_prediction(user_num, business_num, ratings_matrix),2)
    review_edit.loc[record, ['item-item']] = round(item_item_rating_prediction(user_num, business_num, ratings_matrix),2)

review_edit

From the above result, we can see item_based is closer to the real 'stars' most of the time.

def rmse(predictions, targets):
    return np.sqrt(((predictions - targets) ** 2).mean())

√[∑di^2/n]=Re: Reflects the degree of dispersion of a data set.

rmse(review_edit['user-item'], review_edit['stars'])
rmse(review_edit['item-item'], review_edit['stars'])

# Hybrid. Give them different weights.

review_edit.insert(5, 'hybrid1-9', (review_edit['user-item']*.1 + review_edit['item-item']*.9))
review_edit.insert(6, 'hybrid2-8', (review_edit['user-item']*.2 + review_edit['item-item']*.8))
review_edit.insert(7, 'hybrid3-7', (review_edit['user-item']*.3 + review_edit['item-item']*.7))
review_edit.insert(8, 'hybrid4-6', (review_edit['user-item']*.4 + review_edit['item-item']*.6))
review_edit.insert(9, 'hybrid5-5', (review_edit['user-item']*.5 + review_edit['item-item']*.5))
review_edit.insert(10, 'hybrid6-4', (review_edit['user-item']*.6 + review_edit['item-item']*.4))
review_edit.insert(11, 'hybrid7-3', (review_edit['user-item']*.7 + review_edit['item-item']*.3))
review_edit.insert(12, 'hybrid8-2', (review_edit['user-item']*.8 + review_edit['item-item']*.2))
review_edit.insert(13, 'hybrid9-1', (review_edit['user-item']*.9 + review_edit['item-item']*.1))

review_edit

# Print the range of RMSE values for each of the rating prediction columns
# Ranging from using 100% of the item-item ratings to 100% of the user-item ratings
print('item-item:', rmse(review_edit['item-item'], review_edit['stars']))
print('hybrid 1-9:', rmse(review_edit['hybrid1-9'], review_edit['stars']))
print('hybrid 2-8:', rmse(review_edit['hybrid2-8'], review_edit['stars']))
print('hybrid 3-7:', rmse(review_edit['hybrid3-7'], review_edit['stars']))
print('hybrid 4-6:', rmse(review_edit['hybrid4-6'], review_edit['stars']))
print('hybrid 5-5:', rmse(review_edit['hybrid5-5'], review_edit['stars']))
print('hybrid 6-4:', rmse(review_edit['hybrid6-4'], review_edit['stars']))
print('hybrid 7-7:', rmse(review_edit['hybrid7-3'], review_edit['stars']))
print('hybrid 8-8:', rmse(review_edit['hybrid8-2'], review_edit['stars']))
print('hybrid 9-1:', rmse(review_edit['hybrid9-1'], review_edit['stars']))
print('user-item:', rmse(review_edit['user-item'], review_edit['stars']))

Conclusion:
1. Item-based approach is much better than User-based filtering system;
2. Hybrid approach can further reduce our Root Squared Mean Error (RSME);
3. 10% item-based and 90% user-based，gives an answer that is close to the correct answer and minimizes RSME at most time.
