==> 4.Restaurant Recommendation.ipynb <==
# Building the Recommendation System

The purpose of this workbook is to develop a prediction model for any given user.

import numpy as np
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

ratings_matrix = pd.read_pickle('./Dataset/ratings_matrix_new.pkl')

all_business = pd.read_pickle('./Dataset/business.pkl')
unique_business = pd.read_pickle('./Dataset/unique_business.pkl')
unique_user = pd.read_pickle('./Dataset/unique_user.pkl')

all_business.head()

biz_sort = all_business.sort_values('stars', ascending = False)
biz_sort.head(10)

# Review the ratings matrix
ratings_matrix

Same as the fuction we made in "Build Rating Matrix"⬇️

def find_user_similarity(userA, userB, ratings_matrix):
    # Create a True/False list of businesses that were given a rating for each of the two users
    businesses_rated_by_userA = ~ratings_matrix.loc[userA, :].isna()
    businesses_rated_by_userB = ~ratings_matrix.loc[userB, :].isna()

    # Consolidate the two boolean lists into a single one
    businesses_rated_by_AB = businesses_rated_by_userA & businesses_rated_by_userB

    # Capture the rating values of both users for those businesses that were rated by both users
    # Also transform these values into a format suitable for the cosine_similarity function
    ratings_of_userA = ratings_matrix.loc[userA, businesses_rated_by_AB].values.reshape(1, -1)
    ratings_of_userB = ratings_matrix.loc[userB, businesses_rated_by_AB].values.reshape(1, -1)

    # Capture the similaritiy between the two users by comparing their ratings for the set of businesses that they have both provided a rating for
    similarity = cosine_similarity(ratings_of_userA, ratings_of_userB)[0][0]

    return similarity

'''
Create a function to calculate the user-item rating prediction based on cosine similarity, with the following two parameters:
 target_business = business_id value for business for whom rating is being predicted for
 target_user = user_id value for the user for whom rating is being predicted for
'''
def user_item_rating_prediction(target_user, target_business, ratings_matrix):

    similarities_to_target_user = []
    ratings_given_to_target_business = []

    # Create a list of all users that have provided a rating for the target business
    list_of_users_rating_target_business = list(ratings_matrix[~ratings_matrix.iloc[:, target_business].isna()].index)

    # Loop over every user in our target ratings matrix
    # Refer to each user as the 'other_user' since we know that our target user did not provide a rating for our target business and hence is not in this smaller data frame
    for other_user in list_of_users_rating_target_business:
        # To compensate for the value error that may occur when the two users we are comparing have 0 businesses that they have both rated together
        try:
            # Capture the cosine similarity between our target user and the current user from the list of user we are looping over
            similarity = find_user_similarity(target_user, other_user, ratings_matrix)
            similarities_to_target_user.append(similarity)
            ratings_given_to_target_business.append(ratings_matrix.loc[other_user, target_business])
        # If a value error is generated, we simply pass over to the next loop
        # Since we will not be appending no values to neither our list of similarities and list of ratings, we will not be impacting our final calculation
        except:
            pass

    # Calculate the weighted average of all ratings (for those users that have at least 1 business that they have rated together)
    return np.dot(ratings_given_to_target_business, similarities_to_target_user)/np.sum(similarities_to_target_user)

The basic logic is the same as above

def find_business_similarity(businessA, businessB, ratings_matrix):

    users_who_rated_businessA = ~ratings_matrix.loc[:, businessA].isna()
    users_who_rated_businessB = ~ratings_matrix.loc[:, businessB].isna()
    
    users_who_rated_by_AB = users_who_rated_businessA & users_who_rated_businessB
    
    ratings_of_businessA = ratings_matrix.loc[users_who_rated_by_AB, businessA].values.reshape(1, -1)
    ratings_of_businessB = ratings_matrix.loc[users_who_rated_by_AB, businessB].values.reshape(1, -1)
    
    similarity = cosine_similarity(ratings_of_businessA, ratings_of_businessB)[0][0]
    
    return similarity

'''
Create a function to calculate the user-item rating prediction based on cosine similarity, with the following two parameters:
 target_business = business_id value for business for whom rating is being predicted for
 target_user = user_id value for the user for whom rating is being predicted for
'''
def item_item_rating_prediction(target_user, target_business, ratings_matrix):
   
    similarities_to_target_business = []
    ratings_given_by_target_user = []
    
    list_of_businesses_rated_by_target_user = list(ratings_matrix.loc[:, ~ratings_matrix.iloc[target_user, :].isna()].columns)
    
    for other_business in list_of_businesses_rated_by_target_user:
        try:
            similarity = find_business_similarity(target_business, other_business, ratings_matrix)
            similarities_to_target_business.append(similarity)
            ratings_given_by_target_user.append(ratings_matrix.loc[target_user, other_business])
        except:
            pass
    
    return np.dot(ratings_given_by_target_user, similarities_to_target_business)/np.sum(similarities_to_target_business)

Check:

user_item_rating_prediction(22917, 2, ratings_matrix)

item_item_rating_prediction(22917, 2, ratings_matrix)

businesses_in_range = pd.merge(all_business, unique_business, on = 'business_id', how = 'inner')
businesses_in_range.head()

# Make prediction

We use decimal degrees as the standard unit of measurement for latitude and longitude geographic coordinates.

After some research and experimentation, we decided to use a decimal degree distance of 0.015 to determine the range of restaurants nearby to limit the scope of our recommender system. This distance represents a straight distance of around 1.75 km in any one direction.

# Create a set of parameters (to use as an example) 

# default GPS coordinates 
lat_gps = 39.9861726
lon_gps = -75.1322293

# Example user for whom we want to generate ratings for 

ex_user_id = 'hDMM20MqoKVi4VMKamyfqQ'
 
# Determine the unique user number for the selected example user above
ex_user_num = unique_user[unique_user['user_id'] == ex_user_id]
user_num = ex_user_num.iloc[0, 0]

# Set a distance value (in Decimal Degrees)* see explanation above
gps_distance = 0.015

'''
Create a function which outputs a list of restaurants sorted by ratings predicted for a given user

Parameters:
 user_num: as per the matching index from our ratings matrix
 user_lat, user_lon: chosen geographic location to centre our recommendations around
 gps_distance: user inputted distance to determine geographic range, in decimal degrees
'''
def restaurant_predictions(user_num, aim_lat, aim_lon, gps_distance):
    
    # Using the given GPS coordinates and GPS distance range
    min_lat = aim_lat - gps_distance
    max_lat = aim_lat + gps_distance
    min_lon = aim_lon - gps_distance
    max_lon = aim_lon + gps_distance
    
    list_of_businesses_in_range = []
    
    # Looping over each record in our business df, select all the restaurants which fall within this range
    for record in all_business.index:
        
        # Capture the latitude and longitude coordinates for the current business we are looping over
        bus_lat = all_business.iloc[record, 5]
        bus_lon = all_business.iloc[record, 6]
        
        # If the GPS coordinates of the current business falls within our GPS range...
        if (bus_lat > min_lat) and (bus_lat < max_lat) and (bus_lon > min_lon) and (bus_lon < max_lon):
            list_of_businesses_in_range.append(record)
           
    # Given our list of index values for our selected businesses...
    # Create a new df containing only those selected businesses
    businesses_in_range = all_business.loc[list_of_businesses_in_range,:]
    
    # Insert a new column called 'user rating' filled with null values
    businesses_in_range.insert(7, 'user_rating', np.nan)
    
    businesses_in_range = pd.merge(businesses_in_range, unique_business, on = 'business_id', how = 'inner')
    
    # Loop over every record in our df containing the selected restaurants with our user-inputted GPS range
    for record in businesses_in_range.index:
        # Capture the business_num for the current business we are looping over in order to refer to our ratings matrix
        business_num = businesses_in_range.iloc[record, 9]
        # Determine the ratings prediction using the hybrid approach as determined in the previous notebook...
        # Weighted average using 80% of the item-item-rating and 20% of the user-item-rating
        ratings_prediction = item_item_rating_prediction(user_num, business_num, ratings_matrix)*0.9 + user_item_rating_prediction(user_num, business_num, ratings_matrix)*0.1
        # Capture the predicted rating for the given user into our df
        businesses_in_range.iloc[record, 7] = ratings_prediction
    
    # Sort our df containing those selected restaurants within our GPS range by the predicted user rating
    businesses_in_range.sort_values('user_rating', ascending = False, inplace = True)
    
    return businesses_in_range

%%time

recommendations = restaurant_predictions(user_num, lat_gps, lon_gps, gps_distance)

# Show only the top 10 restaurants with the highest predicted user rating for the selected user 
recommendations.head(10)

# delete irrelevant columns and save 'name', 'latitude', 'longitude' and 'star' as our output

import json

res = json.loads(recommendations.to_json())

response = []

biz_list = list(res['name'].keys())

result_limit = 10
for idx in range(result_limit):
    
    biz_idx = biz_list[idx]
    
    biz_info = {}
    
    biz_info["name"] = res["name"][biz_idx]
    
    biz_info["latitude"] = res["latitude"][biz_idx]
    
    biz_info["longtitude"] = res["longitude"][biz_idx]
    
    biz_info["star"] = res["stars"][biz_idx]

    response.append(biz_info)


print(json.dumps(response))
